alert creation logic
