+++ static/app/views/ddm/createAlertModal.tsx
@@ -1,6 +1,5 @@
 import {Fragment, useCallback, useMemo, useState} from 'react';
 import styled from '@emotion/styled';
-import * as qs from 'query-string';
 
 import {ModalRenderProps} from 'sentry/actionCreators/modal';
 import {Button} from 'sentry/components/button';
@@ -18,105 +17,32 @@ import {Tooltip} from 'sentry/components/tooltip';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {PageFilters, Project} from 'sentry/types';
-import {parsePeriodToHours, statsPeriodToDays} from 'sentry/utils/dates';
 import {
   formatMetricUsingFixedUnit,
-  getDDMInterval,
   MetricDisplayType,
   MetricsQuery,
 } from 'sentry/utils/metrics';
+import {formatMRIField, MRIToField, parseMRI} from 'sentry/utils/metrics/mri';
 import {
-  formatMRIField,
-  getUseCaseFromMRI,
-  MRIToField,
-  parseMRI,
-} from 'sentry/utils/metrics/mri';
+  getAlertFormPath,
+  getAlertInterval,
+  getAlertPeriod,
+  getInitialFormState,
+} from 'sentry/utils/metrics/useCreateAlert';
 import {useMetricsData} from 'sentry/utils/metrics/useMetricsData';
 import useOrganization from 'sentry/utils/useOrganization';
 import useProjects from 'sentry/utils/useProjects';
 import useRouter from 'sentry/utils/useRouter';
-import {AVAILABLE_TIME_PERIODS} from 'sentry/views/alerts/rules/metric/triggers/chart';
-import {
-  Dataset,
-  EventTypes,
-  TimePeriod,
-  TimeWindow,
-} from 'sentry/views/alerts/rules/metric/types';
 import {AlertWizardAlertNames} from 'sentry/views/alerts/wizard/options';
 import {getChartSeries} from 'sentry/views/ddm/widget';
-
-interface Props extends ModalRenderProps {
-  metricsQuery: MetricsQuery;
-}
 
 interface FormState {
   environment: string | null;
   project: string | null;
 }
 
-function getInitialFormState(metricsQuery: MetricsQuery): FormState {
-  const project =
-    metricsQuery.projects.length === 1 ? metricsQuery.projects[0].toString() : null;
-  const environment =
-    metricsQuery.environments.length === 1 && project
-      ? metricsQuery.environments[0]
-      : null;
-
-  return {
-    project,
-    environment,
-  };
-}
-
-function getAlertPeriod(metricsQuery: MetricsQuery) {
-  const {period, start, end} = metricsQuery.datetime;
-  const inHours = statsPeriodToDays(period, start, end) * 24;
-
-  switch (true) {
-    case inHours <= 6:
-      return TimePeriod.SIX_HOURS;
-    case inHours <= 24:
-      return TimePeriod.ONE_DAY;
-    case inHours <= 3 * 24:
-      return TimePeriod.THREE_DAYS;
-    case inHours <= 7 * 24:
-      return TimePeriod.SEVEN_DAYS;
-    case inHours <= 14 * 24:
-      return TimePeriod.FOURTEEN_DAYS;
-    default:
-      return TimePeriod.SEVEN_DAYS;
-  }
-}
-
-const TIME_WINDOWS_TO_CHECK = [
-  TimeWindow.ONE_MINUTE,
-  TimeWindow.FIVE_MINUTES,
-  TimeWindow.TEN_MINUTES,
-  TimeWindow.FIFTEEN_MINUTES,
-  TimeWindow.THIRTY_MINUTES,
-  TimeWindow.ONE_HOUR,
-  TimeWindow.TWO_HOURS,
-  TimeWindow.FOUR_HOURS,
-  TimeWindow.ONE_DAY,
-];
-
-function getAlertInterval(metricsQuery, period: TimePeriod) {
-  const useCase = getUseCaseFromMRI(metricsQuery.mri) ?? 'custom';
-  const interval = getDDMInterval(metricsQuery.datetime, useCase);
-  const inMinutes = parsePeriodToHours(interval) * 60;
-
-  function toInterval(timeWindow: TimeWindow) {
-    return `${timeWindow}m`;
-  }
-
-  for (let index = 0; index < TIME_WINDOWS_TO_CHECK.length; index++) {
-    const timeWindow = TIME_WINDOWS_TO_CHECK[index];
-    if (inMinutes <= timeWindow && AVAILABLE_TIME_PERIODS[timeWindow].includes(period)) {
-      return toInterval(timeWindow);
-    }
-  }
-
-  return toInterval(TimeWindow.ONE_HOUR);
+interface Props extends ModalRenderProps {
+  metricsQuery: MetricsQuery;
 }
 
 export function CreateAlertModal({Header, Body, Footer, metricsQuery}: Props) {
@@ -211,29 +137,22 @@ export function CreateAlertModal({Header, Body, Footer, metricsQuery}: Props) {
 
   const handleSubmit = useCallback(() => {
     router.push(
-      `/organizations/${organization.slug}/alerts/new/metric/?${qs.stringify({
-        // Needed, so alerts-create also collects environment via event view
-        createFromDiscover: true,
-        dataset: Dataset.GENERIC_METRICS,
-        eventTypes: EventTypes.TRANSACTION,
+      getAlertFormPath({
+        organizationSlug: organization.slug,
         aggregate: MRIToField(metricsQuery.mri, metricsQuery.op!),
+        query: `${metricsQuery.query} event.type:transaction`.trim(),
         interval: alertInterval,
         statsPeriod: alertPeriod,
-        referrer: 'ddm',
-        // Event type also needs to be added to the query
-        query: `${metricsQuery.query}  event.type:transaction`.trim(),
         environment: formState.environment ?? undefined,
         project: selectedProject!.slug,
-      })}`
+      })
     );
   }, [
+    organization.slug,
+    metricsQuery,
     alertInterval,
     alertPeriod,
     formState.environment,
-    metricsQuery.mri,
-    metricsQuery.op,
-    metricsQuery.query,
-    organization.slug,
     router,
     selectedProject,
   ]);
